refact : PostOrderType 추출 메서드 명확한 책임 분리를 위한 PostUtil 클래스로 이동 #89
