+++ src/main/java/com/midas/shootpointer/domain/post/helper/PostValidationImpl.java
@@ -69,26 +69,4 @@ public class PostValidationImpl implements PostValidation{
         }
     }
 
-    /*==========================
-    *
-    *PostValidationImpl
-    *
-    * @parm type : 조회순/인기순 정렬 type
-    * @return 존재하지 않는 정렬 type일 시 CustomException 발생 / 존재하면 ENUM 클래스 반환.
-    * @author kimdoyeon
-    * @version 1.0.0
-    * @date 25. 9. 10.
-    *
-    ==========================**/
-    @Override
-    public PostOrderType isValidPostOrderType(String type) {
-        PostOrderType postOrderType;
-        try {
-            postOrderType=PostOrderType.valueOf(type);
-        }catch (IllegalArgumentException e){
-            throw new CustomException(ErrorCode.NOT_EXIST_ORDER_TYPE);
-        }
-        return postOrderType;
-    }
-
 }
